requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    nox -s pip_compile
#
asgiref==3.7.2
    # via
    #   -r requirements/prod.txt
    #   django
attrs==23.1.0
    # via
    #   -r requirements/prod.txt
    #   cattrs
cattrs==23.1.2
    # via -r requirements/prod.txt
covdefaults==2.3.0
    # via -r requirements/test.in
coverage==7.3.2
    # via
    #   -r requirements/test.in
    #   covdefaults
django==4.2.6
    # via -r requirements/prod.txt
factory-boy==3.3.0
    # via -r requirements/test.in
faker==19.12.0
    # via factory-boy
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via pytest
pluggy==1.3.0
    # via pytest
pytest==7.4.2
    # via
    #   -r requirements/test.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/prod.txt
    #   django
